Move standard responses in types directory
