+++ app/modules/cinema/endpoints_cinema.py
@@ -7,7 +7,6 @@ from fastapi import Depends, File, HTTPException, UploadFile
 from fastapi.responses import FileResponse
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.core import standard_responses
 from app.core.config import Settings
 from app.core.core_endpoints import models_core
 from app.core.groups.groups_type import AccountType, GroupType
@@ -23,6 +22,7 @@ from app.dependencies import (
     is_user_in,
 )
 from app.modules.cinema import cruds_cinema, schemas_cinema
+from app.types import standard_responses
 from app.types.content_type import ContentType
 from app.types.module import Module
 from app.types.scheduler import Scheduler
