+++ app/modules/phonebook/endpoints_phonebook.py
@@ -5,7 +5,6 @@ from fastapi import Depends, File, HTTPException, UploadFile
 from fastapi.responses import FileResponse
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.core import standard_responses
 from app.core.core_endpoints import models_core
 from app.core.groups import cruds_groups
 from app.core.groups.groups_type import AccountType, GroupType
@@ -18,6 +17,7 @@ from app.dependencies import (
 )
 from app.modules.phonebook import cruds_phonebook, models_phonebook, schemas_phonebook
 from app.modules.phonebook.types_phonebook import RoleTags
+from app.types import standard_responses
 from app.types.content_type import ContentType
 from app.types.module import Module
 from app.utils.tools import (
